+++ tide/master/network/MasterFromWallChannel.h
@@ -70,6 +70,13 @@ signals:
      */
     void receivedRequestFrame( QString uri );
 
+    /**
+     * Emitted after each wall process has rendered a screenshot
+     * @param uri The rendered image
+     * @param source The identifier of the wall process that sent the image
+     */
+    void receivedScreenshot( QImage image, int source );
+
 private:
     Q_DISABLE_COPY( MasterFromWallChannel )